ai_service/requirements.txt
requests
beautifulsoup4
sentence-transformers
scikit-learn
flask
flask-cors
PyMuPDF     # For parsing PDF, extremely high performance
python-docx # For parsing .docx Word documents
playwright